Refactor: Move user-related entities and DTOs to account package
