+++ src/main/java/cz/kocabek/animerecomedationsystem/account/dto/AccWatchlistShowDto.java
@@ -1,11 +1,11 @@
-package cz.kocabek.animerecomedationsystem.user.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+package cz.kocabek.animerecomedationsystem.account.dto;
 
 import java.io.Serializable;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
 /**
- * DTO for {@link cz.kocabek.animerecomedationsystem.user.entity.AccWatchlist}
+ * DTO for {@link cz.kocabek.animerecomedationsystem.account.entity.AccWatchlist}
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record AccWatchlistShowDto(
