--- src/main/java/cz/kocabek/animerecomedationsystem/user/entity/RoleType.java
@@ -1,6 +1,0 @@
-package cz.kocabek.animerecomedationsystem.user.entity;
-
-public enum RoleType {
-    USER,
-    ADMIN
-}
